### 2021-10-19 #3834 @whedon
[PRE REVIEW]: flowTorch - a Python library for analysis and reduced-order modeling of fluid flows
**Submitting author:** @AndreWeiner (<a href="http://orcid.org/0000-0001-5617-1560">Andre Weiner</a>)
**Repository:** <a href="https://github.com/FlowModelingControl/flowtorch" target ="_blank">https://github.com/FlowModelingControl/flowtorch</a>
**Version:** v1.0
