### 2021-10-27 #3860 @whedon
[REVIEW]: flowTorch - a Python library for analysis and reduced-order modeling of fluid flows
**Submitting author:** @AndreWeiner (<a href="http://orcid.org/0000-0001-5617-1560">Andre Weiner</a>)
**Repository:** <a href="https://github.com/FlowModelingControl/flowtorch" target ="_blank">https://github.com/FlowModelingControl/flowtorch</a>
**Version:** v1.0
**Reviewer:** @akaptano, @JaroslavHron, @salrm8, @hkjeldsberg

@akaptano & @JaroslavHron & @salrm8 & @hkjeldsberg, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
