### 2021-10-27 021b1b6
Add requirements.txt file and corresponding description to README; issue #22.
