### 2021-10-27 9cab33a
Remove hard-coded path from PSPExplorer unittest; issue #26.
