### 2021-10-27 83a94df
Add instructions for downloading reduced dataset; issue #26.
