### 2021-10-29 0931524
Add doc strings where missingi; issue #27.
