### 2021-10-29 0608a2a
Add references to compatible Python packages, issue #27.
