### 2021-10-29 0b4a80d
Improve statement of need in README; issue #23.
