### 2021-10-29 0a40712
Revise article according to issue #25.
