+++ paper.bib
@@ -37,6 +37,41 @@
 	journal = {Science Advances}
 }
 
+@article{demo2018,
+  Author = {Demo, Nicola and Tezzele, Marco and Rozza, Gianluigi},
+  Title = {{PyDMD: Python Dynamic Mode Decomposition}},
+  Journal = {The Journal of Open Source Software},
+  Volume = {3},
+  Number = {22},
+  Pages = {530},
+  Year = {2018},
+  Doi = {https://doi.org/10.21105/joss.00530}
+}
+
+@article{desilva2020,
+doi = {10.21105/joss.02104},
+url = {https://doi.org/10.21105/joss.02104},
+year = {2020},
+publisher = {The Open Journal},
+volume = {5},
+number = {49},
+pages = {2104},
+author = {Brian de Silva and Kathleen Champion and Markus Quade and Jean-Christophe Loiseau and J. Kutz and Steven Brunton},
+title = {PySINDy: A Python package for the sparse identification of nonlinear dynamical systems from data},
+journal = {Journal of Open Source Software}
+}
+
+@article{noack2003,
+title={A hierarchy of low-dimensional models for the transient and post-transient cylinder wake},
+volume={497},
+DOI={10.1017/S0022112003006694},
+journal={Journal of Fluid Mechanics},
+publisher={Cambridge University Press},
+author={Noack, Bernd R. and Afanasiev, Konstantin and Morzyński, Marek and Tadmor, Gilead and Thiele, Frank},
+year={2003},
+pages={335–363}
+}
+
 @incollection{paszke2015,
 title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
 author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
