+++ paper.md
@@ -14,7 +14,7 @@ authors:
     orcid: 0000-0002-3219-0545
     affiliation: 1
 affiliations:
- - name: Technische Universit\"at Braunschweig, Institute of Fluid Mechanics, Flow Modeling and Control Group
+ - name: Technische Universität Braunschweig, Institute of Fluid Mechanics, Flow Modeling and Control Group
    index: 1
 date: 05 August 2021
 bibliography: paper.bib
@@ -37,29 +37,31 @@ reduced-order models.
 # Statement of need
 
 Thanks to the increased processing power of modern hardware, fluid flow experiments as well as numerical simulations
-are producing vast amounts of highly resolved, complex data. Those data offer great opportunities to improve complex technical application or to understand natural phenomena. However, gaining insights from the data is becoming increasingly challenging. As data sizes grow, more attention has to be paid to the
-data format. While most researchers prefer simple text-encoded comma-separated value (CSV) files, big datasets require
+are producing vast amounts of highly resolved, complex data. Those data offer great opportunities to optimize industrial processes or to understand natural phenomena. As modern datasets continue to grow, post-processing pipelines will be increasingly important for synthesizing different data formats and facilitating complex data analysis. While most researchers prefer simple text-encoded comma-separated value (CSV) files, big datasets require
 special binary formats, such as [HDF5](https://www.hdfgroup.org/solutions/hdf5/) or [NetCDF](https://en.wikipedia.org/wiki/NetCDF).
 If the data are associated with a structured or an unstructured mesh, VTK files are a popular choice. Other simulation libraries for
-fluid flows, like OpenFOAM, organize mesh and field data in custom folder and file structures. CSV, VTK, or OpenFOAM data may come
+fluid flows, like OpenFOAM, organize mesh and field data in custom folder and file structures. On the experimental side, software packages
+like DaVis allow exporting particle image velocimetry (PIV) snapshots as CSV files. Reading CSV files can be a daunting task, too. A sequence of
+snapshots might be organized in one or multiple files. If the data are stored in a single file, the file must be read first and then the individual snapshots must be extracted following some initially unknown pattern. If the data are spread out over multiple files, the time might be encoded in the file name, but it could be also the case that
+the files are located in individual folders whose names encode the time. The latter structure is typical for OpenFOAM run time post-processing data.
+Moreover, different software packages will create different file headers, which may have to be parsed or sometimes ignored. CSV, VTK, or OpenFOAM data may come
 as binary or text-encoded files. This list is by no means comprehensive in terms of available formats and presents only the tip of the iceberg.
 
 A common research task may be to compare and combine different data sources of the same fluid flow problem for cross-validation
-or to leverage each source's strengths in different kinds of analysis. A typical example would be to compare or combine particle image velocity (PIV) data with sampled
+or to leverage each source's strengths in different kinds of analysis. A typical example would be to compare or combine PIV data with sampled
 planes extracted from a numerical simulation. The simulation offers greater details and additional field information, while the PIV experiment is more trustworthy since it
-is closer to the real application. The PIV data may have to be processed and cleaned before using it in consecutive analysis steps. When confronted with such ideas, researchers will spend a significant amount of time,
-if not most of the time, accessing, converting, and processing the data with different tools and different formats to finally analyze the data in yet another tool.
[…]
