### 2021-10-31 #3860 @hkjeldsberg
Dear all,

The library looks good from my end, and I think @akaptano already addressed the most pressing issues.

From a developer perspective I would recommend collecting the tests in a separate _test_-folder, as this would clean up the current project structure.
Furthermore, a possible improvement is adding some sort of continuous integration, i.e., [TravisCI](https://travis-ci.org/)., [CircleCI](https://circleci.com/), or [AppVeyor](https://www.appveyor.com/) for automated testing.

However, I do not consider these blocking issues, and I am therefore happy to recommend this library for publication.
