### 2021-11-01 75f9112
Add IRQ outlier detection and replacement.
