+++ flowtorch/version.py
@@ -1,3 +1,3 @@
 """flowTorch package version."""
 
-__version__ = "1.0"
+__version__ = "1.1"
