### 2021-11-02 9b0b013
Add robust SVD implementation.
