+++ flowtorch/analysis/svd.py
@@ -1,10 +1,93 @@
 """Classes and functions wrapping around *torch.linalg.svd*.
 """
 
+# standard library packages
+from math import sqrt
+from typing import Tuple, Union
 # third party packages
 import torch as pt
 # flowtorch packages
 from flowtorch.data.utils import format_byte_size
+
+
+def inexact_alm_matrix_complection(data_matrix: pt.Tensor, sparsity: float = 1.0,
+                                   tol: float = 1.0e-6, max_iter: int = 100,
+                                   verbose: bool = False) -> Tuple[pt.Tensor, pt.Tensor]:
+    """Split a data matrix in low rank and sparse contributions.
+
+    This function implements the *inexact augmented Lagrange multiplier
+    matrix completion* algorithm to solve the *principal component pursuit*
+    problem. The implementation is based on the Matlab code of Isabel Scherl
+    (link_), which is in turn based on the LRSLibrary_.
+
+    .. _link: https://github.com/ischerl/RPCA-PIV/blob/master/functions/inexact_alm_rpca.m
+    .. _LRSLibrary: https://github.com/andrewssobral/lrslibrary
+
+    :param data_matrix: input data matrix; snapshots must be
+        organized as column vectors
+    :type data_matrix: pt.Tensor
+    :param sparsity: factor to compute Lagrangian multiplyer for sparsity
+        (typically named *lambda*); lower values lead to more agressive
+        filtering
+    :type sparsity: float, optional
+    :param tol: tolerance for the normalized Frobenius norm of the difference
+        between original data and the sum of low rank and sparse contributions;
+        defaults to 1.0e-6
+    :type tol: float, optional
+    :param max_iter: maximum number of iteration before to give up, defaults to 100
+    :type max_iter: int, optional
+    :param verbose: residual is printed for every iteration if True;
+        defaults to False
+    :type verbose: bool, optional
+    :return: tuple holding the low rank and the sparse matrices
+    :rtype: Tuple[pt.Tensor, pt.Tensor]
+    """
+    row, col = data_matrix.shape
+    lambda_0 = sparsity / sqrt(row)
+    # low rank and sparse matices
+    L, S = pt.zeros_like(data_matrix), pt.zeros_like(data_matrix)
+    # matrix of Lagrange multipliers
+    Y = data_matrix.detach().clone()
+    norm_two = pt.linalg.svdvals(Y)[0].item()
+    norm_inf = pt.linalg.norm(Y, float("inf")).item()
+    dual_norm = max(norm_two, norm_inf)
+    Y /= dual_norm
+    # more hyperparameters
+    mu = 1.25 / norm_two
+    norm_data = pt.linalg.norm(data_matrix)
+    sv = 10
+    rho = 1.5
+
+    for i in range(max_iter):
+        temp = data_matrix - L + Y/mu
+        S = pt.maximum(temp - lambda_0/mu, pt.tensor(0.0))
+        S += pt.minimum(temp + lambda_0/mu, pt.tensor(0.0))
+        U, s, VH = pt.linalg.svd(data_matrix - S + Y/mu, full_matrices=False)
+        # truncate SVD
+        svp = s[s > 1.0/mu].shape[0]
+        if svp < sv:
+            sv = min(svp+1, col)
+        else:
+            sv = min(svp + round(0.05*col), col)
+        L = U[:, :svp] @ pt.diag(s[:svp] - 1.0/mu) @ VH[:svp, :]
+        # print(L[0,:])
+        # Z is the residual matrix
+        Z = data_matrix - L - S
+        # update Lagrange multipliers
+        Y += mu*Z
+        mu = min(mu*rho, mu*1.0e7)
+        # check convergence
+        residual = pt.linalg.norm(Z) / norm_data
+        if residual < tol:
[…]
