### 2021-11-02 577fafa
Add options for robust SVD to DMD.
