+++ flowtorch/analysis/dmd.py
@@ -2,7 +2,7 @@
 """
 
 # standard library packages
-from typing import Tuple, Set
+from typing import Tuple, Set, Union
 # third party packages
 import torch as pt
 from numpy import pi
@@ -32,10 +32,13 @@ class DMD(object):
     tensor([-2.3842e-06, -4.2345e+01, -1.8552e+01])
     >>> dmd.amplitude
     tensor([10.5635+0.j, -0.0616+0.j, -0.0537+0.j])
+    >>> dmd = DMD(data_matrix, dt=0.1, rank=3, robust=True)
+    >>> dmd = DMD(data_matrix, dt=0.1, rank=3, robust={"tol": 1.0e-5, "verbose" : True})
 
     """
 
-    def __init__(self, data_matrix: pt.Tensor, dt: float, rank: int = None):
+    def __init__(self, data_matrix: pt.Tensor, dt: float, rank: int = None,
+                 robust: Union[bool, dict] = False):
         """Create DMD instance based on data matrix and time step. 
 
         :param data_matrix: data matrix whose columns are formed by the individual snapshots
@@ -44,10 +47,14 @@ class DMD(object):
         :type dt: float
         :param rank: rank for SVD truncation, defaults to None
         :type rank: int, optional
+        :param robust: data_matrix is split into low rank and sparse contributions
+            if True or if dictionary with options for Inexact ALM algorithm; the SVD
+            is computed only on the low rank matrix
+        :type robust: Union[bool,dict]
         """
         self._dm = data_matrix
         self._dt = dt
-        self._svd = SVD(self._dm[:, :-1], rank)
+        self._svd = SVD(self._dm[:, :-1], rank, robust)
         self._eigvals, self._eigvecs, self._modes = self._compute_mode_decomposition()
 
     def _compute_mode_decomposition(self):
