### 2021-11-04 6256459
Fix CNM prediction if enocder is None.
