### 2021-11-08 #3860 @JaroslavHron
Hi, i like the paper/library. Especially the the documentation (https://flowmodelingcontrol.github.io/flowtorch-docs/1.0/) is nicely written.

Apart of the already resolved issues pointed out by @akaptano, I have no further remarks and I would recommend this  for publication as it is now.

May be a small one:
* from a user point of view I might be tempted to look in the documentation for some hints on writing custom data reader to import flow data in some other format. For example I might have computed flow data in the XDMF format  -  probably this might be easily derived from the available hdf5, vtk and openfoam readers.
