### 2021-11-09 #3860 @AndreWeiner
Hi @JaroslavHron,
thank you very much for your feedback. I already thought about the idea of adding also more advanced programming tutorials to the documentation (there is something like that already for simple linear algebra and tensor operations with PyTorch). The data loaders are probably the first structure one might want to extend.
Best, Andre
