### 2021-11-11 e620a2b
Print statistics on iqr outlier replacement.
