### 2021-11-22 #3860 @AndreWeiner
Dear @Kevin-Mattheus-Moerman,
a quick question: I guess the DOI printed in the draft is not going to be the final DOI but acts as a placeholder, right?
I'll have to submit another article next week, in which I'd like to cite flowTorch using a proper DOI.
Thanks,
Andre
