### 2021-12-03 #3860 @salrm8
Congratulations @AndreWeiner for the nice software and especially the great documentation.
The paper is written well, and improved by comments from @akaptano.

The only suggestions (minor) are:

- Creating a folder for the paper,
- Possibly adding all tests into one folder "tests".
- A guide how to adjust data from other flow solvers (dataloader).

But even without these, I strongly recommend accepting flowTorch!
