### 2021-12-03 #3860 @AndreWeiner
Dear @akaptano, @JaroslavHron, @hkjeldsberg, @salrm8, and @Kevin-Mattheus-Moerman,
before this issue gets closed, I would like to take the time to thank all of you for your time and your contributions to improve flowTorch. Honest word, this has been the most constructive review process I've ever gone through. Absolutely no comparison to the process in classical journals.
Thanks again,
Andre
