### 2021-12-03 #3860 @salrm8
> Hi @salrm8, first of all, thanks for your feedback. I made a copy of the listed suggestions. I think you still have to edit the checkboxes in the first post to complete your review. Cheers

Hi @AndreWeiner , @Kevin-Mattheus-Moerman ,
Unfortunately I cannot edit the checkboxes, although I am logged-in and have accepted the invitation
