### 2021-12-05 #3860 @arfon
> @arfon can you comment on the DOI question ☝️, is the DOI listed "reserved" for this work? So that if accepted this becomes the actual DOI right?

The listed DOI in the draft is the DOI that this submission will eventually have yes.
