### 2021-12-06 #3860 @Kevin-Mattheus-Moerman
@AndreWeiner this looks ready to proceed to acceptance, congratulations. I proofread the paper too and it mostly looks fine. There are a few minor steps before we proceed:

- [ ] Minor comment on paper. You refer to `Jupyter labs`, did you mean to refer to `Jupyter notebooks`? Please check. You can update the paper here if needed.
- [ ] Once you are happy with the paper, can you archive a copy of the software repository on ZENODO? Some do this manually while others prefer these automated steps: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content.
- [ ] Once the ZENODO archive is created please ensure the meta-data like the title, and author set matches that of the paper. You can also add ORCID id's for authors where possible.
- [ ] Can you confirm the version label for the published software, is it still at `v1.0`?

Thanks, let me know if you have any questions.
