### 2021-12-09 #3860 @AndreWeiner
Dear @Kevin-Mattheus-Moerman

- I changed labs to notebooks
- a snapshot of v1.0 is archived on Zenodo; [link](https://zenodo.org/record/5770244#.YbIPGLso9hE)
- software name, authors, orcids, etc. are the same for archive and article
- yes, the version is still *v1.0*

Best, Andre
