### 2021-12-12 #3860 @Kevin-Mattheus-Moerman
@AndreWeiner one issue remains the DOI of the archived version should not match that of the paper. It looks like you made the Zenodo archive point to the JOSS DOI, please remove that and assign a separate ZENODO DOI there.
