### 2021-12-12 #3860 @Kevin-Mattheus-Moerman
@openjournals/dev it looks the authors have assigned the JOSS DOI to their ZENODO archive. I'm recommending the above. Let me know if anything else (or a different approach) is needed to fix this.
