### 2021-12-12 #3860 @AndreWeiner
@Kevin-Mattheus-Moerman,
sorry, that was a misunderstanding on my side. I got in touch with the Zenodo support staff to chance the DOI. I'll let you know once the issue is resolved.
Best, Andre
