### 2021-12-13 #3860 @AndreWeiner
Alright, the DOI issue has been fixed. Here is the link again to the updated Zenodo record: [click](https://zenodo.org/record/5770244)
Sorry for the trouble @Kevin-Mattheus-Moerman; hope it works now.
Best, Andre
