### 2021-12-13 #3860 @Kevin-Mattheus-Moerman
@openjournals/dev I just accepted this paper but accidentally did so before fixing an invalid DOI, can this still be fixed?
```
https://doi.org/10.21105/joss.00530 is INVALID because of 'https://doi.org/' prefix
```
