### 2021-12-13 256907d
Add reference to flowTorch article to README.
