+++ paper.bib
@@ -45,7 +45,7 @@
   Number = {22},
   Pages = {530},
   Year = {2018},
-  Doi = {https://doi.org/10.21105/joss.00530}
+  Doi = {10.21105/joss.00530}
 }
 
 @article{desilva2020,
